refactor_code_for_repeated_functions
