+++ vllm/model_executor/kernels/linear/mixed_precision/cpu.py
@@ -5,6 +5,7 @@ import torch
 
 from vllm import _custom_ops as ops
 from vllm.model_executor.layers.quantization.utils.quant_utils import (
+    get_isa_hint,
     pack_quantized_values_into_int32,
     unpack_quantized_values_into_int32,
 )
@@ -68,7 +69,7 @@ class CPUWNA16LinearKernel(MPLinearKernel):
         p_w_k, p_w_n = packed_weight.size()
         input_size = p_w_k * pack_factor
         output_size = p_w_n
-        isa_hint = _get_isa_hint(layer.scales.dtype)
+        isa_hint = get_isa_hint(layer.scales.dtype)
         layer.isa_hint = isa_hint
 
         layer.qzeros = None
@@ -116,11 +117,3 @@ class CPUWNA16LinearKernel(MPLinearKernel):
             isa_hint=layer.isa_hint,
         )
         return x
-
-
-def _get_isa_hint(dtype: torch.dtype) -> str:
-    supports_amx = torch._C._cpu._is_amx_tile_supported()
-    if supports_amx and dtype in (torch.bfloat16,):
-        return "amx"
-    else:
-        return "vec"
