+++ vllm/model_executor/models/paddleocr_vl.py
@@ -80,6 +80,7 @@ from .utils import (
     AutoWeightsLoader,
     PPMissingLayer,
     WeightsMapper,
+    all_gather_interleave,
     is_pp_missing_parameter,
     maybe_prefix,
 )
@@ -505,25 +506,6 @@ class SiglipVisionEmbeddings(nn.Module):
                 "Unsupported pixel_values dimension:"
                 f" {pixel_values.dim()}. Expected 4 or 5."
             )
-
-
-def all_gather_interleave(local_tensor: torch.Tensor, hidden_size: int, tp_size: int):
-    """All-gather the input tensor interleavely across model parallel group."""
-    import torch.distributed as dist
-
-    gathered_tensors = [torch.zeros_like(local_tensor) for _ in range(tp_size)]
-    dist.all_gather(
-        gathered_tensors, local_tensor, group=parallel_state.get_tp_group().device_group
-    )
-
-    gathered_tensors_split = [
-        torch.split(tensor, hidden_size // tp_size, -1) for tensor in gathered_tensors
-    ]
-    ordered_tensors = [
-        tensor for pair in zip(*gathered_tensors_split) for tensor in pair
-    ]
-    result_tensor = torch.cat(ordered_tensors, dim=-1)
-    return result_tensor
 
 
 class SiglipAttention(nn.Module):
